content-based-movie-recommendation/requirements.txt
blinker==1.6.3
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
Flask==3.0.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
load-dotenv==0.1.0
MarkupSafe==2.1.3
numpy==1.26.0
pandas==1.5.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
tmdbv3api==1.9.0
tzdata==2023.3
urllib3==2.0.6
Werkzeug==3.0.0
